setを実装する

+++ tests/vector/VectorBenchmark.cpp
@@ -1,6 +1,6 @@
 #include "vector/VectorBenchmark.hpp"
 
-const int VectorBenchmark::loop_count = 10000;
+const size_t VectorBenchmark::loop_count = 10000;
 
 VectorBenchmark::VectorBenchmark() : timer(Timer()) {}
 
@@ -16,7 +16,7 @@ void VectorBenchmark::test_all()
 void VectorBenchmark::test_clear()
 {
     timer.start();
-    for (int i = 0; i < loop_count; i++)
+    for (size_t i = 0; i < loop_count; i++)
     {
         ft::vector<int> vec(100);
         vec.clear();
@@ -30,7 +30,7 @@ void VectorBenchmark::test_insert()
     ft::vector<int> vec;
 
     timer.start();
-    for (int i = 0; i < loop_count; i++)
+    for (size_t i = 0; i < loop_count; i++)
     {
         vec.insert(vec.end(), 42);
     }
@@ -43,7 +43,7 @@ void VectorBenchmark::test_erase()
     ft::vector<int> vec(loop_count);
 
     timer.start();
-    for (int i = 0; i < loop_count; i++)
+    for (size_t i = 0; i < loop_count; i++)
     {
         vec.erase(vec.end() - 1);
     }
@@ -56,7 +56,7 @@ void VectorBenchmark::test_push_back()
     ft::vector<int> vec;
 
     timer.start();
-    for (int i = 0; i < loop_count; i++)
+    for (size_t i = 0; i < loop_count; i++)
     {
         vec.push_back(i);
     }
@@ -69,7 +69,7 @@ void VectorBenchmark::test_pop_back()
     ft::vector<int> vec(loop_count);
 
     timer.start();
-    for (int i = 0; i < loop_count; i++)
+    for (size_t i = 0; i < loop_count; i++)
     {
         vec.pop_back();
     }
